+++ test/Resources/barManager.js
@@ -1,0 +1,10 @@
+export default class BarManager {
+
+  constructor (fooManager) {
+    this._fooManager = fooManager
+  }
+
+  get fooManager () {
+    return this._fooManager
+  }
+}