+++ test/Resources/fooManager.js
@@ -1,0 +1,6 @@
+export default class FooManager {
+
+  constructor () {
+    return FooManager.prototype.fooManagerCalls++
+  }
+}